requirements.txt
streamlit>=1.37.0
openai>=1.30.0
python-dotenv>=1.0.1
pdfminer.six>=20220524
docx2txt>=0.8
pandas>=2.2.0

